# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "description": "Keeping track of your habits and moods",
    "version": "1.0.0",
    "title": "Mood and Habit Tracker"
  },
  "host": "habit-mood-tracker.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "moods",
      "description": "Operations for moods"
    },
    {
      "name": "user",
      "description": "Operations for users"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/moods": {
      "get": {
        "tags": [
          "moods"
        ],
        "summary": "Get all moods",
        "operationId": "getAllMoods",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/moods"
              }
            }
          },
          "500": {
            "description": "Error occurred"
          }
        }
      }
    },
    "/moods/{moodName}": {
      "get": {
        "tags": [
          "moods"
        ],
        "summary": "Get specific mood",
        "operationId": "getMood",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "moodName",
            "in": "path",
            "description": "Name of the mood",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/moods"
              }
            }
          },
          "400": {
            "description": "Invalid Mood Id"
          },
          "404": {
            "description": "Mood not found"
          },
          "500": {
            "description": "Other Error Caused"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get all users",
        "operationId": "getAllUsers",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/users"
              }
            }
          },
          "500": {
            "description": "Error occurred"
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Creates User",
        "description": "Creates a new user",
        "operationId": "createNewUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create user object",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/users"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/users"
              }
            }
          },
          "400": {
            "description": "Error occurred"
          }
        }
      }
    },
    "/users/{username}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get specific user",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username you are looking for",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/users"
              }
            }
          },
          "400": {
            "description": "Invalid User Id"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update a user",
        "description": "Add the username and body for the user",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username you are wanting to update",
            "required": true,
            "type": "string"
          },
          {
          "name": "body",
          "in": "body",
          "description": "Things you want to update in the user",
          "required": true,
          "type": "string",
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/users"
            }
          }
        }
        ],
        "responses": {
          "204": {
            "description": "Successfully modified user"
          },
          "400": {
            "description": "Invalid Username"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete a user",
        "description": "Add the username of the user you want to delete",
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username you are wanting to update",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully modified user"
          },
          "400": {
            "description": "Invalid Username"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "definitions": {
    "users": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "habitTracker": {
          "$ref": "#/definitions/users_habitTracker"
        },
        "chosenMoods": {
          "$ref": "#/definitions/users_chosenMoods"
        }
      }
    },
    "moods": {
      "type": "object",
      "properties": {
        "moodName": {
          "type": "string"
        },
        "moodDescription": {
          "type": "string"
        },
        "moodColor": {
          "type": "string"
        },
        "moodEmoji": {
          "type": "string"
        }
      }
    },
    "users_habitTracker_habits": {
      "type": "object",
      "properties": {
        "habitName": {
          "type": "string"
        },
        "build": {
          "type": "boolean"
        },
        "quit": {
          "type": "boolean"
        },
        "daily": {
          "type": "boolean"
        },
        "weekly": {
          "type": "boolean"
        },
        "monthly": {
          "type": "boolean"
        },
        "yearly": {
          "type": "boolean"
        },
        "timesPerGoalPeriod": {
          "type": "integer",
          "format": "int32"
        },
        "countDone": {
          "type": "integer",
          "format": "int32"
        },
        "backgroundColor": {
          "type": "string"
        },
        "daysCompleted": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "date"
          }
        }
      }
    },
    "users_habitTracker": {
      "type": "object",
      "properties": {
        "habits": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/users_habitTracker_habits"
          }
        }
      }
    },
    "users_chosenMoods_moods": {
      "type": "object",
      "properties": {
        "day": {
          "type": "string",
          "format": "date"
        },
        "moodName": {
          "type": "string"
        }
      }
    },
    "users_chosenMoods": {
      "type": "object",
      "properties": {
        "moods": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/users_chosenMoods_moods"
          }
        }
      }
    }
  }
}